dex-trading-capstone-starter/requirements.txt
pandas
numpy
httpx
python-dotenv
solana
matplotlib
